segp_oral_cancer_backend/requirements.txt
torch>=2.2
torchvision>=0.17
scikit-learn>=1.4
openslide-python>=1.3.1
opencv-python-headless>=4.9
Pillow>=10.3
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
albumentations>=1.4
pyyaml>=6.0
fastapi>=0.111
uvicorn[standard]>=0.30
tensorboard>=2.16
python-multipart>=0.0.9
